Enable support for JS and CSS
For some bizarre reason PHPCS supports sending CSS and JS files on to
other external tools. It has been requested that this linter support this,
even though it is a much better idea to use the Linter providers dedicated
to those tools. Setting this up to be tested is rather difficult, so no
specs are provided.

Re-implementation of #113.