Error: Invalid point encountered - consistently with PHPCS 1.x (#234)
With PHPCS 1.x all of the line numbers are off by one, which consistently causes invalid point encountered, especially for long lines. Seems the filename was being specified in STDIN but only PHP 2.0, <2.6 supports this, so for PHPCS 1.x it just throws off the line number by 1 for every single error.