Handle all tested versions in the specs
Update the specs to handle the differences in the expected results based
on the version that should be running in the CI environment. If
PHPCS_VER is not specified it defaults to the expected values for the
latest version (currently 3.0.0). Also greatly enhances the tab width
specs to account for several cases that weren't being tested before.