Fix handling of forced standards with default tabWidth
If the tabWidth setting was left at the default value, a standard that
forces the tab width value was in use, and a version of PHPCS was in use
that _doesn't_ ignore that setting then the code previously would assume
that it didn't need to correct the value coming from PHPCS when it
actually did.

This bug was introduced in #260 due to a disconnect between the default
value of 1 and the function assuming a default of 0.